refactor(test): split prf-handler/key-cache specs; exclude test-helpers from coverage
- prf-handler.spec.ts (678 lines) を passkey / secret の 2 ファイルへ分割し、
  共有モックを prf-handler.test-helpers.ts へ抽出。
- key-cache.spec.ts (502 lines) を core / memory の 2 ファイルへ分割
  （セットアップが軽量なためヘルパーは設けず scaffold を各ファイルに保持）。
- *.test-helpers.ts は実コードではないため、両 package の vitest coverage
  exclude に追加（カバレッジ分母から除外）。

テスト件数は SDK 373 / iframe 149 で不変。
