+++ packages/nosskey-sdk/vitest.config.ts
@@ -8,7 +8,13 @@ export default defineConfig({
       reporter: ['text', 'json', 'json-summary', 'html', 'lcov'],
       reportsDirectory: './coverage',
       include: ['src/**/*.ts'],
-      exclude: ['src/**/*.{test,spec}.ts', 'src/test-utils.ts', 'src/index.ts', 'src/types.ts'],
+      exclude: [
+        'src/**/*.{test,spec}.ts',
+        'src/**/*.test-helpers.ts',
+        'src/test-utils.ts',
+        'src/index.ts',
+        'src/types.ts',
+      ],
       reportOnFailure: true,
     },
   },
